refactor(viewer): isolate agent display models
