+++ crates/oasis7_viewer/software_safe_src/agent_context_visual_fixture.js
@@ -1,4 +1,4 @@
-import { AGENT_INTENT_SUMMARIES } from "./agent_intent_surface.jsx";
+import { AGENT_INTENT_SUMMARIES } from "./agent_intent_display_model.js";
 
 export const AGENT_CONTEXT_FIXTURE_MODES = Object.freeze(["rich", "unavailable"]);
 export const AGENT_CONTEXT_FIXTURE_STATES = Object.freeze(["current", "stale", "reconnecting"]);
