+++ crates/oasis7_viewer/software_safe_src/agent_intent_visual_fixture.js
@@ -157,4 +157,4 @@ export function installAgentIntentV2VisualFixture(
     core.requestRender();
   };
 }
-import { AGENT_INTENT_SUMMARIES } from "./agent_intent_surface.jsx";
+import { AGENT_INTENT_SUMMARIES } from "./agent_intent_display_model.js";
