+++ crates/oasis7_viewer/viewer.js
@@ -15873,7 +15873,6 @@ function installBranchCommitmentVisualFixture(fixtures, { core: core2, setFixtur
     setFixturePlayerAuth2();
   };
 }
-var _tmpl$$6 = /* @__PURE__ */ template(`<span class="badge badge--accent">`), _tmpl$2$6 = /* @__PURE__ */ template(`<div class=agent-intent__status-row>`), _tmpl$3$5 = /* @__PURE__ */ template(`<div class=agent-intent__summary>`), _tmpl$4$4 = /* @__PURE__ */ template(`<div class="agent-intent__detail agent-intent__receipt"><span class=metric__label>`), _tmpl$5$4 = /* @__PURE__ */ template(`<div class=agent-intent__detail><span class=metric__label></span><span class=agent-intent__summary>`), _tmpl$6$3 = /* @__PURE__ */ template(`<div class="agent-intent__detail agent-intent__lifecycle">`), _tmpl$7$3 = /* @__PURE__ */ template(`<div class="agent-intent__detail agent-intent__next-step"><span class=metric__label></span><span class=agent-intent__summary>`), _tmpl$8$3 = /* @__PURE__ */ template(`<section class=agent-intent aria-live=polite><div class="agent-intent__heading metric__label"></div><div class=agent-intent__state>`);
 const PLAYER_SAFE_COPY_SCHEMA_VERSION = 1;
 const INTENT_STATUS_LABELS = {
   proposed: ["已提出", "Proposed"],
@@ -15951,7 +15950,7 @@ function hasReceiptReference(receiptRef, intent) {
   if (textValue$2(receiptRef.intent_id) !== textValue$2(intent?.intent_id) || textValue$2(receiptRef.world_id) !== textValue$2(intent?.world_id) || receiptEventId === null || receiptEventId === "0") return false;
   return counterIdentity(receiptRef.reorg_epoch) === counterIdentity(intent?.reorg_epoch) && counterIdentity(receiptRef.logical_time) === counterIdentity(intent?.logical_time) && counterIdentity(receiptRef.event_seq) === counterIdentity(intent?.event_seq);
 }
-function copy$1(locale, key) {
+function agentIntentCopy(locale, key) {
   const zh = String(locale || "").toLowerCase().startsWith("zh");
   const values = {
     heading: ["当前意图", "Current Intent"],
@@ -15979,157 +15978,71 @@ function statusLabel(locale, status) {
   return values ? values[String(locale || "").toLowerCase().startsWith("zh") ? 0 : 1] : "";
 }
 function unavailable(locale, nextStep = NEXT_STEP_ALLOWLIST.unavailable, extra = {}) {
-  return {
-    kind: "unavailable",
-    label: copy$1(locale, "unavailable"),
-    nextStep,
-    receiptState: "not_applicable",
-    ...extra
-  };
+  return { kind: "unavailable", label: agentIntentCopy(locale, "unavailable"), nextStep, receiptState: "not_applicable", ...extra };
 }
 function copyVersion(intent) {
   const explicit = COPY_KEY_FIELDS.map((field) => intent[field]).find((value2) => value2 !== void 0 && value2 !== null);
   return explicit === void 0 ? PLAYER_SAFE_COPY_SCHEMA_VERSION : explicit;
 }
 function allowlistedIntentCopy(intent, status) {
-  if (copyVersion(intent) !== PLAYER_SAFE_COPY_SCHEMA_VERSION) return {
-    valid: false,
-    value: ""
-  };
+  if (copyVersion(intent) !== PLAYER_SAFE_COPY_SCHEMA_VERSION) return { valid: false, value: "" };
   const expected = AGENT_INTENT_SUMMARIES[status];
   const key = textValue$2(intent.summary_key || intent.summaryKey);
-  if (key && key !== status) return {
-    valid: false,
-    value: ""
-  };
+  if (key && key !== status) return { valid: false, value: "" };
   const supplied = textValue$2(intent.summary ?? intent.message);
-  if (!supplied || supplied !== expected) return {
-    valid: false,
-    value: ""
-  };
-  return {
-    valid: true,
-    value: expected
-  };
+  if (!supplied || supplied !== expected) return { valid: false, value: "" };
+  return { valid: true, value: expected };
 }
 function allowlistedReason(intent, status) {
   const key = textValue$2(intent.reason_code || intent.reason_key || intent.reasonKey).toLowerCase();
   const supplied = textValue$2(intent.reason_summary);
-  if (!key) return supplied ? {
-    valid: false,
-    label: "",
-    summary: ""
-  } : {
-    valid: true,
-    label: "",
-    summary: ""
-  };
-  if (!Object.prototype.hasOwnProperty.call(REASON_ALLOWLIST, key)) return {
-    valid: false,
-    label: "",
-    summary: ""
-  };
+  if (!key) return supplied ? { valid: false, label: "", summary: "" } : { valid: true, label: "", summary: "" };
+  if (!Object.prototype.hasOwnProperty.call(REASON_ALLOWLIST, key)) return { valid: false, label: "", summary: "" };
   const declaredKey = textValue$2(intent.reason_key || intent.reasonKey).toLowerCase();
-  if (declaredKey && declaredKey !== key) return {
-    valid: false,
-    label: "",
-    summary: ""
-  };
-  if (supplied && supplied !== REASON_ALLOWLIST[key]) return {
-    valid: false,
-    label: "",
-    summary: ""
-  };
-  if (!TERMINAL_INTENT_STATUSES.has(status) && status !== "blocked") return {
-    valid: true,
-    label: "",
-    summary: ""
-  };
-  return {
-    valid: true,
-    label: key,
-    summary: REASON_ALLOWLIST[key]
-  };
+  if (declaredKey && declaredKey !== key) return { valid: false, label: "", summary: "" };
+  if (supplied && supplied !== REASON_ALLOWLIST[key]) return { valid: false, label: "", summary: "" };
+  if (!TERMINAL_INTENT_STATUSES.has(status) && status !== "blocked") return { valid: true, label: "", summary: "" };
+  return { valid: true, label: key, summary: REASON_ALLOWLIST[key] };
 }
 function allowlistedNextStep(intent, status, stateKind) {
   const declared = textValue$2(intent.next_step_key || intent.nextStepKey).toLowerCase();
-  if (declared && !Object.prototype.hasOwnProperty.call(NEXT_STEP_ALLOWLIST, declared)) return {
-    valid: false,
-    value: ""
-  };
+  if (declared && !Object.prototype.hasOwnProperty.call(NEXT_STEP_ALLOWLIST, declared)) return { valid: false, value: "" };
   const fallbackKey = stateKind === "current" && (status === "blocked" || status === "rejected") ? status : stateKind;
   const expected = NEXT_STEP_ALLOWLIST[declared || fallbackKey] || "";
   const supplied = textValue$2(intent.next_step || intent.next_step_hint);
-  if (supplied && supplied !== expected) return {
-    valid: false,
-    value: ""
-  };
-  return {
-    valid: true,
-    value: expected
-  };
+  if (supplied && supplied !== expected) return { valid: false, value: "" };
+  return { valid: true, value: expected };
 }
 function describeAgentIntent(intent, locale = "en", connectionStatus = "connected") {
   if (!intent || typeof intent !== "object") return unavailable(locale);
   if (intent.schema_version !== 2 || textValue$2(intent.source_class) !== "runtime_projection") return unavailable(locale);
   const connection = textValue$2(connectionStatus).toLowerCase();
-  if (connection === "connecting" || connection === "reconnecting") return {
-    kind: "reconnecting",
-    label: copy$1(locale, "reconnecting"),
-    nextStep: NEXT_STEP_ALLOWLIST.reconnecting,
-    receiptState: "not_applicable"
-  };
-  if (connection && connection !== "connected") return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, {
-    label: copy$1(locale, "offline")
-  });
+  if (connection === "connecting" || connection === "reconnecting") return { kind: "reconnecting", label: agentIntentCopy(locale, "reconnecting"), nextStep: NEXT_STEP_ALLOWLIST.reconnecting, receiptState: "not_applicable" };
+  if (connection && connection !== "connected") return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, { label: agentIntentCopy(locale, "offline") });
   const controlState = textValue$2(intent.control_state).toLowerCase();
   if (!ALLOWED_CONTROL_STATES.has(controlState)) return unavailable(locale);
-  if (controlState === "control_lost") return {
-    kind: controlState,
-    label: copy$1(locale, "hiddenControlLost"),
-    nextStep: NEXT_STEP_ALLOWLIST.control_lost,
-    receiptState: "hidden"
-  };
-  if (controlState === "read_only") return {
-    kind: controlState,
-    label: copy$1(locale, "hiddenReadOnly"),
-    nextStep: NEXT_STEP_ALLOWLIST.read_only,
-    receiptState: "hidden"
-  };
-  if (controlState === "unauthorized") return {
-    kind: controlState,
-    label: copy$1(locale, "hiddenUnauthorized"),
-    nextStep: NEXT_STEP_ALLOWLIST.unauthorized,
-    receiptState: "hidden"
-  };
+  if (controlState === "control_lost") return { kind: controlState, label: agentIntentCopy(locale, "hiddenControlLost"), nextStep: NEXT_STEP_ALLOWLIST.control_lost, receiptState: "hidden" };
+  if (controlState === "read_only") return { kind: controlState, label: agentIntentCopy(locale, "hiddenReadOnly"), nextStep: NEXT_STEP_ALLOWLIST.read_only, receiptState: "hidden" };
+  if (controlState === "unauthorized") return { kind: controlState, label: agentIntentCopy(locale, "hiddenUnauthorized"), nextStep: NEXT_STEP_ALLOWLIST.unauthorized, receiptState: "hidden" };
   if (controlState === "unavailable" || !textValue$2(intent.intent_id) || !hasAuthoritativePosition(intent)) return unavailable(locale);
   const status = textValue$2(intent.status).toLowerCase();
   if (!statusLabel(locale, status)) return unavailable(locale);
   const freshness = textValue$2(intent.freshness).toLowerCase();
   if (!ALLOWED_FRESHNESS.has(freshness)) return unavailable(locale);
   const receiptState = status === "completed" ? hasReceiptReference(intent.receipt_ref, intent) ? "confirmed" : "missing" : "not_applicable";
-  const receiptLabel = receiptState === "confirmed" ? copy$1(locale, "receipt") : receiptState === "missing" ? copy$1(locale, "receiptMissing") : "";
-  if (receiptState === "missing") return unavailable(locale, NEXT_STEP_ALLOWLIST.missing_receipt, {
-    receiptState,
-    receiptLabel
-  });
+  const receiptLabel = receiptState === "confirmed" ? agentIntentCopy(locale, "receipt") : receiptState === "missing" ? agentIntentCopy(locale, "receiptMissing") : "";
+  if (receiptState === "missing") return unavailable(locale, NEXT_STEP_ALLOWLIST.missing_receipt, { receiptState, receiptLabel });
   const safeCopy = allowlistedIntentCopy(intent, status);
-  if (!safeCopy.valid) return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, {
-    receiptState
-  });
+  if (!safeCopy.valid) return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, { receiptState });
   const reason = allowlistedReason(intent, status);
-  if (!reason.valid) return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, {
-    receiptState
-  });
+  if (!reason.valid) return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, { receiptState });
   const stateKind = freshness === "stale" ? "stale" : freshness === "reconnecting" ? "reconnecting" : freshness === "conflict" ? "conflict" : "current";
   const nextStep = allowlistedNextStep(intent, status, stateKind);
-  if (!nextStep.valid) return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, {
-    receiptState
-  });
-  const lifecycleNote = intent.duplicate === true || intent.replayed === true || intent.replay === true ? copy$1(locale, "replayed") : textValue$2(intent.replaced_by) ? copy$1(locale, "replaced") : "";
+  if (!nextStep.valid) return unavailable(locale, NEXT_STEP_ALLOWLIST.unavailable, { receiptState });
+  const lifecycleNote = intent.duplicate === true || intent.replayed === true || intent.replay === true ? agentIntentCopy(locale, "replayed") : textValue$2(intent.replaced_by) ? agentIntentCopy(locale, "replaced") : "";
   const base = {
     kind: stateKind,
-    label: stateKind === "stale" ? copy$1(locale, "stale") : stateKind === "conflict" ? copy$1(locale, "needsConfirmation") : stateKind === "reconnecting" ? copy$1(locale, "reconnecting") : copy$1(locale, "current"),
+    label: stateKind === "stale" ? agentIntentCopy(locale, "stale") : stateKind === "conflict" ? agentIntentCopy(locale, "needsConfirmation") : stateKind === "reconnecting" ? agentIntentCopy(locale, "reconnecting") : agentIntentCopy(locale, "current"),
     statusLabel: stateKind === "stale" ? "" : statusLabel(locale, status),
     message: safeCopy.value,
     receiptState,
@@ -16140,106 +16053,6 @@ function describeAgentIntent(intent, locale = "en", connectionStatus = "connecte
     nextStep: nextStep.value
   };
   return base;
-}
-function AgentIntentSurface(props) {
-  const locale = () => props.locale || "en";
-  const model = () => describeAgentIntent(props.intent, locale(), props.connectionStatus);
-  const hidden = () => ["control_lost", "read_only", "unauthorized", "unavailable"].includes(model().kind);
-  const showReceiptConfirmation = () => props.showReceiptConfirmation !== false;
-  return (() => {
-    var _el$ = _tmpl$8$3(), _el$2 = _el$.firstChild, _el$3 = _el$2.nextSibling;
-    insert(_el$2, () => copy$1(locale(), "heading"));
-    insert(_el$3, () => model().label);
-    insert(_el$, createComponent(Show, {
-      get when() {
-        return !hidden();
-      },
-      get children() {
-        return [(() => {
-          var _el$4 = _tmpl$2$6();
-          insert(_el$4, createComponent(Show, {
-            get when() {
-              return model().statusLabel;
-            },
-            get children() {
-              var _el$5 = _tmpl$$6();
-              insert(_el$5, () => model().statusLabel);
-              return _el$5;
-            }
-          }));
-          return _el$4;
-        })(), createComponent(Show, {
-          get when() {
-            return model().message;
-          },
-          get children() {
-            var _el$6 = _tmpl$3$5();
-            insert(_el$6, () => model().message);
-            return _el$6;
-          }
-        }), createComponent(Show, {
-          get when() {
-            return memo(() => !!showReceiptConfirmation())() && model().receiptLabel;
-          },
-          get children() {
-            var _el$7 = _tmpl$4$4(), _el$8 = _el$7.firstChild;
-            insert(_el$8, () => model().receiptLabel);
-            return _el$7;
-          }
-        }), createComponent(Show, {
-          get when() {
-            return model().reasonLabel || model().reasonSummary;
-          },
-          get children() {
-            var _el$9 = _tmpl$5$4(), _el$0 = _el$9.firstChild, _el$1 = _el$0.nextSibling;
-            insert(_el$0, () => copy$1(locale(), "reason"));
-            insert(_el$1, () => model().reasonSummary);
-            return _el$9;
-          }
-        }), createComponent(Show, {
-          get when() {
-            return model().lifecycleNote;
-          },
-          get children() {
-            var _el$10 = _tmpl$6$3();
-            insert(_el$10, () => model().lifecycleNote);
-            return _el$10;
-          }
-        })];
-      }
-    }), null);
-    insert(_el$, createComponent(Show, {
-      get when() {
-        return memo(() => !!(showReceiptConfirmation() && hidden()))() && model().receiptLabel;
-      },
-      get children() {
-        var _el$11 = _tmpl$4$4(), _el$12 = _el$11.firstChild;
-        insert(_el$12, () => model().receiptLabel);
-        return _el$11;
-      }
-    }), null);
-    insert(_el$, createComponent(Show, {
-      get when() {
-        return model().nextStep;
-      },
-      get children() {
-        var _el$13 = _tmpl$7$3(), _el$14 = _el$13.firstChild, _el$15 = _el$14.nextSibling;
-        insert(_el$14, () => copy$1(locale(), "nextStep"));
-        insert(_el$15, () => model().nextStep);
-        return _el$13;
-      }
-    }), null);
-    createRenderEffect((_p$) => {
-      var _v$ = model().kind, _v$2 = model().receiptState || "not_applicable";
-      _v$ !== _p$.e && setAttribute(_el$, "data-agent-intent-state", _p$.e = _v$);
-      _v$2 !== _p$.t && setAttribute(_el$, "data-agent-intent-receipt-state", _p$.t = _v$2);
-      return _p$;
-    }, {
-      e: void 0,
-      t: void 0
-    });
-    return _el$;
-  })();
 }
 const AGENT_INTENT_STATUSES = [
   "proposed",
@@ -16566,7 +16379,7 @@ function installMajorWorldEventCrisisVisualFixture(fixtures, { core: core2, view
     core2.requestRender();
   };
 }
-var _tmpl$$5 = /* @__PURE__ */ template(`<button data-testid=viewer-available-action-reprioritize>`), _tmpl$2$5 = /* @__PURE__ */ template(`<div class=toolbar data-testid=viewer-reprioritize-action>`), _tmpl$3$4 = /* @__PURE__ */ template(`<div id=viewer-reprioritize-status role=alert tabindex=-1 class=feedback-detail>`), _tmpl$4$3 = /* @__PURE__ */ template(`<div id=viewer-reprioritize-status aria-live=polite class=feedback-detail>`), _tmpl$5$3 = /* @__PURE__ */ template(`<form><label for=viewer-reprioritize-goal></label><textarea id=viewer-reprioritize-goal rows=3 aria-describedby="viewer-reprioritize-help viewer-reprioritize-status"></textarea><div id=viewer-reprioritize-help class=feedback-detail></div><div class=toolbar><button type=button></button><button type=submit>`);
+var _tmpl$$6 = /* @__PURE__ */ template(`<button data-testid=viewer-available-action-reprioritize>`), _tmpl$2$6 = /* @__PURE__ */ template(`<div class=toolbar data-testid=viewer-reprioritize-action>`), _tmpl$3$5 = /* @__PURE__ */ template(`<div id=viewer-reprioritize-status role=alert tabindex=-1 class=feedback-detail>`), _tmpl$4$4 = /* @__PURE__ */ template(`<div id=viewer-reprioritize-status aria-live=polite class=feedback-detail>`), _tmpl$5$4 = /* @__PURE__ */ template(`<form><label for=viewer-reprioritize-goal></label><textarea id=viewer-reprioritize-goal rows=3 aria-describedby="viewer-reprioritize-help viewer-reprioritize-status"></textarea><div id=viewer-reprioritize-help class=feedback-detail></div><div class=toolbar><button type=button></button><button type=submit>`);
 function ReprioritizeActionForm(props) {
   const [open, setOpen] = createSignal(false);
   const [draft, setDraft] = createSignal("");
@@ -16625,14 +16438,14 @@ function ReprioritizeActionForm(props) {
     setSubmitted(true);
   };
   return (() => {
-    var _el$ = _tmpl$2$5();
+    var _el$ = _tmpl$2$6();
     insert(_el$, createComponent(Show, {
       get when() {
         return !open();
       },
       get fallback() {
         return (() => {
-          var _el$3 = _tmpl$5$3(), _el$4 = _el$3.firstChild, _el$5 = _el$4.nextSibling, _el$6 = _el$5.nextSibling, _el$9 = _el$6.nextSibling, _el$0 = _el$9.firstChild, _el$1 = _el$0.nextSibling;
+          var _el$3 = _tmpl$5$4(), _el$4 = _el$3.firstChild, _el$5 = _el$4.nextSibling, _el$6 = _el$5.nextSibling, _el$9 = _el$6.nextSibling, _el$0 = _el$9.firstChild, _el$1 = _el$0.nextSibling;
           _el$3.$$keydown = (event) => {
             if (event.key === "Escape" && !inFlight()) {
               event.preventDefault();
@@ -16653,7 +16466,7 @@ function ReprioritizeActionForm(props) {
               return localError();
             },
             get children() {
-              var _el$7 = _tmpl$3$4();
+              var _el$7 = _tmpl$3$5();
               var _ref$2 = errorNode;
               typeof _ref$2 === "function" ? use(_ref$2, _el$7) : errorNode = _el$7;
               insert(_el$7, localError);
@@ -16665,7 +16478,7 @@ function ReprioritizeActionForm(props) {
               return memo(() => !!!localError())() && inFlight();
             },
             get children() {
-              var _el$8 = _tmpl$4$3();
+              var _el$8 = _tmpl$4$4();
               insert(_el$8, () => props.tr(props.locale, "正在认证并提交新目标…", "Authenticating and submitting the new goal…"));
               return _el$8;
             }
@@ -16687,7 +16500,7 @@ function ReprioritizeActionForm(props) {
         })();
       },
       get children() {
-        var _el$2 = _tmpl$$5();
+        var _el$2 = _tmpl$$6();
         _el$2.$$click = () => {
           setOpen(true);
           queueMicrotask(() => textarea?.focus());
@@ -16846,7 +16659,7 @@ function createViewerAgentClaimDisplayModel({ state: state2, tr: tr2 }) {
   }
   return { agentBindingForId: agentBindingForId2, agentClaimUsesCurrentBoundAgent, buildAgentClaimAction: buildAgentClaimAction2, buildAgentClaimTargets: buildAgentClaimTargets2, describeAgentSessionStatus: describeAgentSessionStatus2, hasAgentClaimSessionBoundary: hasAgentClaimSessionBoundary2, hasExecutableAgentClaim: hasExecutableAgentClaim2, normalizedId: normalizedId2, publishedClaimChoiceCandidates: publishedClaimChoiceCandidates2, slot1ClaimChoiceNeedsDefer, slot1ClaimChoiceQuote: slot1ClaimChoiceQuote2 };
 }
-var _tmpl$$4 = /* @__PURE__ */ template(`<div class=event-list>`), _tmpl$2$4 = /* @__PURE__ */ template(`<div class=feedback-detail><strong></strong>: `), _tmpl$3$3 = /* @__PURE__ */ template(`<div class=event-card data-testid=claim-choice-rationale><div class=event-card__title><span>`), _tmpl$4$2 = /* @__PURE__ */ template(`<div class=event-card><div class=event-card__title><span></span><span class="badge badge--warn"></span></div><div class=feedback-detail>`), _tmpl$5$2 = /* @__PURE__ */ template(`<span class="badge badge--warn">`), _tmpl$6$2 = /* @__PURE__ */ template(`<span class=badge>`), _tmpl$7$2 = /* @__PURE__ */ template(`<div class=badge-row>`), _tmpl$8$2 = /* @__PURE__ */ template(`<div class=event-card__meta>`), _tmpl$9$2 = /* @__PURE__ */ template(`<div class=feedback-detail>`), _tmpl$0$2 = /* @__PURE__ */ template(`<div class=event-card><div class=event-card__title><span></span><span class="badge badge--accent">`);
+var _tmpl$$5 = /* @__PURE__ */ template(`<div class=event-list>`), _tmpl$2$5 = /* @__PURE__ */ template(`<div class=feedback-detail><strong></strong>: `), _tmpl$3$4 = /* @__PURE__ */ template(`<div class=event-card data-testid=claim-choice-rationale><div class=event-card__title><span>`), _tmpl$4$3 = /* @__PURE__ */ template(`<div class=event-card><div class=event-card__title><span></span><span class="badge badge--warn"></span></div><div class=feedback-detail>`), _tmpl$5$3 = /* @__PURE__ */ template(`<span class="badge badge--warn">`), _tmpl$6$3 = /* @__PURE__ */ template(`<span class=badge>`), _tmpl$7$3 = /* @__PURE__ */ template(`<div class=badge-row>`), _tmpl$8$3 = /* @__PURE__ */ template(`<div class=event-card__meta>`), _tmpl$9$2 = /* @__PURE__ */ template(`<div class=feedback-detail>`), _tmpl$0$2 = /* @__PURE__ */ template(`<div class=event-card><div class=event-card__title><span></span><span class="badge badge--accent">`);
 function AgentClaimChoiceCard(props) {
   const publishedCandidates = () => props.publishedCandidates?.() || [];
   const choiceQuote = () => props.choiceQuote?.() || null;
@@ -16878,7 +16691,7 @@ function AgentClaimChoiceCard(props) {
       return publishedCandidates().length > 0;
     },
     get children() {
-      var _el$ = _tmpl$$4();
+      var _el$ = _tmpl$$5();
       insert(_el$, createComponent(For, {
         get each() {
           return publishedCandidates();
@@ -16892,7 +16705,7 @@ function AgentClaimChoiceCard(props) {
               return candidate.name;
             },
             get children() {
-              var _el$30 = _tmpl$8$2();
+              var _el$30 = _tmpl$8$3();
               insert(_el$30, () => candidate.name);
               return _el$30;
             }
@@ -16947,14 +16760,14 @@ function AgentClaimChoiceCard(props) {
       return hasPublishedRationale();
     },
     get children() {
-      var _el$2 = _tmpl$3$3(), _el$3 = _el$2.firstChild, _el$4 = _el$3.firstChild;
+      var _el$2 = _tmpl$3$4(), _el$3 = _el$2.firstChild, _el$4 = _el$3.firstChild;
       insert(_el$4, () => tr2("候选路线理由", "Candidate route rationale"));
       insert(_el$2, createComponent(Show, {
         get when() {
           return startingLocation();
         },
         get children() {
-          var _el$5 = _tmpl$2$4(), _el$6 = _el$5.firstChild;
+          var _el$5 = _tmpl$2$5(), _el$6 = _el$5.firstChild;
           _el$6.nextSibling;
           insert(_el$6, () => tr2("起始位置", "Starting location"));
           insert(_el$5, startingLocation, null);
@@ -16966,7 +16779,7 @@ function AgentClaimChoiceCard(props) {
           return specialtySummary();
         },
         get children() {
-          var _el$8 = _tmpl$2$4(), _el$9 = _el$8.firstChild;
+          var _el$8 = _tmpl$2$5(), _el$9 = _el$8.firstChild;
           _el$9.nextSibling;
           insert(_el$9, () => tr2("专长 / 能力", "Specialty / capabilities"));
           insert(_el$8, specialtySummary, null);
@@ -16978,7 +16791,7 @@ function AgentClaimChoiceCard(props) {
           return firstIndustrialGoalHelp();
         },
         get children() {
-          var _el$1 = _tmpl$2$4(), _el$10 = _el$1.firstChild;
+          var _el$1 = _tmpl$2$5(), _el$10 = _el$1.firstChild;
           _el$10.nextSibling;
           insert(_el$10, () => tr2("首个工业目标帮助", "First industrial goal help"));
           insert(_el$1, firstIndustrialGoalHelp, null);
@@ -16990,7 +16803,7 @@ function AgentClaimChoiceCard(props) {
           return riskSummary();
         },
         get children() {
-          var _el$12 = _tmpl$2$4(), _el$13 = _el$12.firstChild;
+          var _el$12 = _tmpl$2$5(), _el$13 = _el$12.firstChild;
           _el$13.nextSibling;
           insert(_el$13, () => tr2("候选风险", "Candidate risk"));
           insert(_el$12, riskSummary, null);
@@ -17002,7 +16815,7 @@ function AgentClaimChoiceCard(props) {
           return recommendationReason();
         },
         get children() {
-          var _el$15 = _tmpl$2$4(), _el$16 = _el$15.firstChild;
+          var _el$15 = _tmpl$2$5(), _el$16 = _el$15.firstChild;
           _el$16.nextSibling;
           insert(_el$16, () => tr2("推荐理由", "Recommendation reason"));
           insert(_el$15, recommendationReason, null);
@@ -17016,7 +16829,7 @@ function AgentClaimChoiceCard(props) {
       return isRationaleMissingDefer();
     },
     get children() {
-      var _el$18 = _tmpl$4$2(), _el$19 = _el$18.firstChild, _el$20 = _el$19.firstChild, _el$21 = _el$20.nextSibling, _el$22 = _el$19.nextSibling;
+      var _el$18 = _tmpl$4$3(), _el$19 = _el$18.firstChild, _el$20 = _el$19.firstChild, _el$21 = _el$20.nextSibling, _el$22 = _el$19.nextSibling;
       insert(_el$20, () => tr2("暂不确认", "Wait before confirming"));
       insert(_el$21, () => tr2("暂缓", "Defer"));
       insert(_el$22, () => tr2(`当前可支付 ${upfrontAmount()} upfront，但确认后只能维持 ${upkeepRunway()} 个完整 upkeep epoch。尚未发布 canonical 路线理由，因此不推荐任何候选。请在理由发布且有额外可用于 upkeep 的 eligible balance 后再评估；仅补足资金不等于被推荐。`, `The ${upfrontAmount()} upfront cost is payable now, but confirmation leaves ${upkeepRunway()} full upkeep epochs. No canonical route rationale is published, so no candidate is recommended. Reassess after a rationale is published and you have additional eligible upkeep balance; funding alone does not make a candidate recommended.`));
@@ -17027,13 +16840,13 @@ function AgentClaimChoiceCard(props) {
       return memo(() => !!(fallbackLabel() || choiceClassLabel()))() && !isRationaleMissingDefer();
     },
     get children() {
-      var _el$23 = _tmpl$7$2();
+      var _el$23 = _tmpl$7$3();
       insert(_el$23, createComponent(Show, {
         get when() {
           return fallbackLabel();
         },
         get children() {
-          var _el$24 = _tmpl$5$2();
+          var _el$24 = _tmpl$5$3();
           insert(_el$24, fallbackLabel);
           return _el$24;
         }
@@ -17043,7 +16856,7 @@ function AgentClaimChoiceCard(props) {
           return choiceClassLabel();
         },
         get children() {
-          var _el$25 = _tmpl$6$2();
+          var _el$25 = _tmpl$6$3();
           insert(_el$25, choiceClassLabel);
           return _el$25;
         }
@@ -17052,20 +16865,20 @@ function AgentClaimChoiceCard(props) {
     }
   })];
 }
-var _tmpl$$3 = /* @__PURE__ */ template(`<div class=feedback-detail><div class=metric__label>`), _tmpl$2$3 = /* @__PURE__ */ template(`<div class="feedback-detail first-delivery-preview"><div class=metric__label>`), _tmpl$3$2 = /* @__PURE__ */ template(`<div>`);
+var _tmpl$$4 = /* @__PURE__ */ template(`<div class=feedback-detail><div class=metric__label>`), _tmpl$2$4 = /* @__PURE__ */ template(`<div class="feedback-detail first-delivery-preview"><div class=metric__label>`), _tmpl$3$3 = /* @__PURE__ */ template(`<div>`);
 function FirstDeliveryPreview(props) {
   const preview = () => props.preview || {};
   const locale = () => props.locale;
   const tr2 = props.tr;
   return (() => {
-    var _el$ = _tmpl$2$3(), _el$2 = _el$.firstChild;
+    var _el$ = _tmpl$2$4(), _el$2 = _el$.firstChild;
     insert(_el$2, () => tr2(locale(), "首单交付预览", "First delivery preview"));
     insert(_el$, createComponent(Show, {
       get when() {
         return preview().localNeed;
       },
       get children() {
-        var _el$3 = _tmpl$$3(), _el$4 = _el$3.firstChild;
+        var _el$3 = _tmpl$$4(), _el$4 = _el$3.firstChild;
         insert(_el$4, () => tr2(locale(), "本地需求", "Local need"));
         insert(_el$3, () => preview().localNeed, null);
         return _el$3;
@@ -17076,7 +16889,7 @@ function FirstDeliveryPreview(props) {
         return preview().expectedOutput;
       },
       get children() {
-        var _el$5 = _tmpl$$3(), _el$6 = _el$5.firstChild;
+        var _el$5 = _tmpl$$4(), _el$6 = _el$5.firstChild;
         insert(_el$6, () => tr2(locale(), "预计产出", "Expected output"));
         insert(_el$5, () => preview().expectedOutput, null);
         return _el$5;
@@ -17087,14 +16900,14 @@ function FirstDeliveryPreview(props) {
         return preview().requiredInputs.length > 0;
       },
       get children() {
-        var _el$7 = _tmpl$$3(), _el$8 = _el$7.firstChild;
+        var _el$7 = _tmpl$$4(), _el$8 = _el$7.firstChild;
         insert(_el$8, () => tr2(locale(), "所需输入", "Required inputs"));
         insert(_el$7, createComponent(For, {
           get each() {
             return preview().requiredInputs;
           },
           children: (input) => (() => {
-            var _el$13 = _tmpl$3$2();
+            var _el$13 = _tmpl$3$3();
             insert(_el$13, input);
             return _el$13;
           })()
@@ -17107,7 +16920,7 @@ function FirstDeliveryPreview(props) {
         return preview().valueTiming;
       },
       get children() {
-        var _el$9 = _tmpl$$3(), _el$0 = _el$9.firstChild;
+        var _el$9 = _tmpl$$4(), _el$0 = _el$9.firstChild;
         insert(_el$0, () => tr2(locale(), "价值时机", "Value timing"));
         insert(_el$9, () => preview().valueTiming, null);
         return _el$9;
@@ -17118,7 +16931,7 @@ function FirstDeliveryPreview(props) {
         return preview().leverageClassUnlocked;
       },
       get children() {
-        var _el$1 = _tmpl$$3(), _el$10 = _el$1.firstChild;
+        var _el$1 = _tmpl$$4(), _el$10 = _el$1.firstChild;
         insert(_el$10, () => tr2(locale(), "解锁杠杆", "Leverage unlocked"));
         insert(_el$1, () => preview().leverageClassUnlocked, null);
         return _el$1;
@@ -17129,7 +16942,7 @@ function FirstDeliveryPreview(props) {
         return preview().returnVisitHook;
       },
       get children() {
-        var _el$11 = _tmpl$$3(), _el$12 = _el$11.firstChild;
+        var _el$11 = _tmpl$$4(), _el$12 = _el$11.firstChild;
         insert(_el$12, () => tr2(locale(), "回访钩子", "Return visit hook"));
         insert(_el$11, () => preview().returnVisitHook, null);
         return _el$11;
@@ -17197,7 +17010,6 @@ function recoveryOptionVisualFixture() {
     }
   ];
 }
-var _tmpl$$2 = /* @__PURE__ */ template(`<div class=agent-activity__field><span class=metric__label></span><span>`), _tmpl$2$2 = /* @__PURE__ */ template(`<div class=agent-activity><div class="agent-activity__heading metric__label"></div><div class=agent-activity__state>`);
 const KNOWN_ACTIVITY_STATUSES = /* @__PURE__ */ new Set(["idle", "executing", "blocked", "waiting", "unavailable"]);
 function textValue$1(value2) {
   if (value2 === null || value2 === void 0) {
@@ -17272,11 +17084,12 @@ function describeAgentActivity(activity, locale = "en") {
     reason
   };
 }
+var _tmpl$$3 = /* @__PURE__ */ template(`<div class=agent-activity__field><span class=metric__label></span><span>`), _tmpl$2$3 = /* @__PURE__ */ template(`<div class=agent-activity><div class="agent-activity__heading metric__label"></div><div class=agent-activity__state>`);
 function AgentActivitySurface(props) {
   const locale = () => props.locale || "en";
   const model = () => describeAgentActivity(props.activity, locale());
   return (() => {
-    var _el$ = _tmpl$2$2(), _el$2 = _el$.firstChild, _el$3 = _el$2.nextSibling;
+    var _el$ = _tmpl$2$3(), _el$2 = _el$.firstChild, _el$3 = _el$2.nextSibling;
     insert(_el$2, () => activityCopy(locale(), "currentActivity"));
     insert(_el$3, () => model().label);
     insert(_el$, createComponent(Show, {
@@ -17284,7 +17097,7 @@ function AgentActivitySurface(props) {
         return memo(() => model().kind === "blocked")() && model().operation;
       },
       get children() {
-        var _el$4 = _tmpl$$2(), _el$5 = _el$4.firstChild, _el$6 = _el$5.nextSibling;
+        var _el$4 = _tmpl$$3(), _el$5 = _el$4.firstChild, _el$6 = _el$5.nextSibling;
         insert(_el$5, () => activityCopy(locale(), "operation"));
         insert(_el$6, () => model().operation);
         return _el$4;
@@ -17295,7 +17108,7 @@ function AgentActivitySurface(props) {
         return memo(() => !!(model().kind !== "unavailable" && model().kind !== "idle" && model().kind !== "unavailable"))() && model().targetLabel;
       },
       get children() {
-        var _el$7 = _tmpl$$2(), _el$8 = _el$7.firstChild, _el$9 = _el$8.nextSibling;
+        var _el$7 = _tmpl$$3(), _el$8 = _el$7.firstChild, _el$9 = _el$8.nextSibling;
         insert(_el$8, () => activityCopy(locale(), "target"));
         insert(_el$9, () => model().targetLabel);
         return _el$7;
@@ -17306,7 +17119,7 @@ function AgentActivitySurface(props) {
         return memo(() => !!(model().kind !== "unavailable" && model().kind !== "idle" && !model().targetLabel))() && model().operation;
       },
       get children() {
-        var _el$0 = _tmpl$$2(), _el$1 = _el$0.firstChild, _el$10 = _el$1.nextSibling;
+        var _el$0 = _tmpl$$3(), _el$1 = _el$0.firstChild, _el$10 = _el$1.nextSibling;
         insert(_el$1, () => activityCopy(locale(), "target"));
         insert(_el$10, () => activityCopy(locale(), "targetUnavailable"));
         return _el$0;
@@ -17317,13 +17130,114 @@ function AgentActivitySurface(props) {
         return memo(() => model().kind === "blocked")() && model().reason;
       },
       get children() {
-        var _el$11 = _tmpl$$2(), _el$12 = _el$11.firstChild, _el$13 = _el$12.nextSibling;
+        var _el$11 = _tmpl$$3(), _el$12 = _el$11.firstChild, _el$13 = _el$12.nextSibling;
         insert(_el$12, () => activityCopy(locale(), "reason"));
         insert(_el$13, () => model().reason);
         return _el$11;
       }
     }), null);
     createRenderEffect(() => setAttribute(_el$, "data-agent-activity-state", model().kind));
+    return _el$;
+  })();
+}
+var _tmpl$$2 = /* @__PURE__ */ template(`<span class="badge badge--accent">`), _tmpl$2$2 = /* @__PURE__ */ template(`<div class=agent-intent__status-row>`), _tmpl$3$2 = /* @__PURE__ */ template(`<div class=agent-intent__summary>`), _tmpl$4$2 = /* @__PURE__ */ template(`<div class="agent-intent__detail agent-intent__receipt"><span class=metric__label>`), _tmpl$5$2 = /* @__PURE__ */ template(`<div class=agent-intent__detail><span class=metric__label></span><span class=agent-intent__summary>`), _tmpl$6$2 = /* @__PURE__ */ template(`<div class="agent-intent__detail agent-intent__lifecycle">`), _tmpl$7$2 = /* @__PURE__ */ template(`<div class="agent-intent__detail agent-intent__next-step"><span class=metric__label></span><span class=agent-intent__summary>`), _tmpl$8$2 = /* @__PURE__ */ template(`<section class=agent-intent aria-live=polite><div class="agent-intent__heading metric__label"></div><div class=agent-intent__state>`);
+function AgentIntentSurface(props) {
+  const locale = () => props.locale || "en";
+  const model = () => describeAgentIntent(props.intent, locale(), props.connectionStatus);
+  const hidden = () => ["control_lost", "read_only", "unauthorized", "unavailable"].includes(model().kind);
+  const showReceiptConfirmation = () => props.showReceiptConfirmation !== false;
+  return (() => {
+    var _el$ = _tmpl$8$2(), _el$2 = _el$.firstChild, _el$3 = _el$2.nextSibling;
+    insert(_el$2, () => agentIntentCopy(locale(), "heading"));
+    insert(_el$3, () => model().label);
+    insert(_el$, createComponent(Show, {
+      get when() {
+        return !hidden();
+      },
+      get children() {
+        return [(() => {
+          var _el$4 = _tmpl$2$2();
+          insert(_el$4, createComponent(Show, {
+            get when() {
+              return model().statusLabel;
+            },
+            get children() {
+              var _el$5 = _tmpl$$2();
+              insert(_el$5, () => model().statusLabel);
+              return _el$5;
+            }
+          }));
+          return _el$4;
+        })(), createComponent(Show, {
+          get when() {
+            return model().message;
+          },
+          get children() {
+            var _el$6 = _tmpl$3$2();
+            insert(_el$6, () => model().message);
+            return _el$6;
+          }
+        }), createComponent(Show, {
+          get when() {
+            return memo(() => !!showReceiptConfirmation())() && model().receiptLabel;
+          },
+          get children() {
+            var _el$7 = _tmpl$4$2(), _el$8 = _el$7.firstChild;
+            insert(_el$8, () => model().receiptLabel);
+            return _el$7;
+          }
+        }), createComponent(Show, {
+          get when() {
+            return model().reasonLabel || model().reasonSummary;
+          },
+          get children() {
+            var _el$9 = _tmpl$5$2(), _el$0 = _el$9.firstChild, _el$1 = _el$0.nextSibling;
+            insert(_el$0, () => agentIntentCopy(locale(), "reason"));
+            insert(_el$1, () => model().reasonSummary);
+            return _el$9;
+          }
+        }), createComponent(Show, {
+          get when() {
+            return model().lifecycleNote;
+          },
+          get children() {
+            var _el$10 = _tmpl$6$2();
+            insert(_el$10, () => model().lifecycleNote);
+            return _el$10;
+          }
+        })];
+      }
+    }), null);
+    insert(_el$, createComponent(Show, {
+      get when() {
+        return memo(() => !!(showReceiptConfirmation() && hidden()))() && model().receiptLabel;
+      },
+      get children() {
+        var _el$11 = _tmpl$4$2(), _el$12 = _el$11.firstChild;
+        insert(_el$12, () => model().receiptLabel);
+        return _el$11;
+      }
+    }), null);
+    insert(_el$, createComponent(Show, {
+      get when() {
+        return model().nextStep;
+      },
+      get children() {
+        var _el$13 = _tmpl$7$2(), _el$14 = _el$13.firstChild, _el$15 = _el$14.nextSibling;
+        insert(_el$14, () => agentIntentCopy(locale(), "nextStep"));
+        insert(_el$15, () => model().nextStep);
+        return _el$13;
+      }
+    }), null);
+    createRenderEffect((_p$) => {
+      var _v$ = model().kind, _v$2 = model().receiptState || "not_applicable";
+      _v$ !== _p$.e && setAttribute(_el$, "data-agent-intent-state", _p$.e = _v$);
+      _v$2 !== _p$.t && setAttribute(_el$, "data-agent-intent-receipt-state", _p$.t = _v$2);
+      return _p$;
+    }, {
+      e: void 0,
+      t: void 0
+    });
     return _el$;
   })();
 }
